requirements/base.txt
celery
certifi
Django==2
django-annoying
django-appconf
django-autoslug
django-constance
django-filter
django-fontawesome
django-redis
django-rest-swagger
django-fontawesome
djangorestframework
djangorestframework-camel-case
google-api-python-client
httplib2
ipython
ipython-genutils
itypes
jedi
kombu
lxml
Markdown
MarkupSafe
mimesis
model-mommy
oauth2client
oauthlib
openapi-codec
pafy
parso
pexpect
pickleshare
Pillow
prompt-toolkit
psycopg2
ptyprocess
Pygments
PyJWT
pytils
pytz
PyYAML
redis
requests
requests-oauthlib
simplegeneric
simplejson
social-auth-app-django
social-auth-core
django-ckeditor
django-admin-easy
sorl-thumbnail
django-select2
openpyxl
django-ajax-selects
django-compressor
drf-renderer-xlsx
django_compressor
django-nested-inline
django-debug-toolbar
django-picklefield